- Bumped version to 2.0.0-SNAPSHOT - Slightly reorganized source tree
