+++ pom.xml
@@ -4,7 +4,7 @@
 
 	<groupId>org.digitalmediaserver</groupId>
 	<artifactId>crowdin-maven-plugin</artifactId>
-	<version>1.1.3-SNAPSHOT</version>
+	<version>2.0.0-SNAPSHOT</version>
 	<packaging>maven-plugin</packaging>
 	<prerequisites>
 		<maven>3.2.3</maven>
