+++ src/main/java/org/digitalmediaserver/crowdin/PushCrowdinMojo.java
@@ -30,10 +30,10 @@ import org.apache.http.entity.mime.content.AbstractContentBody;
 import org.apache.http.entity.mime.content.FileBody;
 import org.apache.http.entity.mime.content.InputStreamBody;
 import org.apache.maven.plugin.MojoExecutionException;
-import org.digitalmediaserver.crowdin.configuration.FileType;
+import org.digitalmediaserver.crowdin.api.CrowdinAPI;
+import org.digitalmediaserver.crowdin.api.FileType;
 import org.digitalmediaserver.crowdin.configuration.UpdateOption;
 import org.digitalmediaserver.crowdin.configuration.TranslationFileSet;
-import org.digitalmediaserver.crowdin.tool.CrowdinAPI;
 import org.digitalmediaserver.crowdin.tool.NSISUtil;
 import org.jdom2.Document;
 import org.jdom2.Element;
