+++ src/main/java/org/digitalmediaserver/crowdin/api/FileType.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
-package org.digitalmediaserver.crowdin.configuration;
+package org.digitalmediaserver.crowdin.api;
 
 import static org.digitalmediaserver.crowdin.AbstractCrowdinMojo.isBlank;
 import java.util.Locale;
