+++ src/main/java/org/digitalmediaserver/crowdin/configuration/AbstractFileSet.java
@@ -24,6 +24,8 @@ import java.util.List;
 import java.util.Properties;
 import javax.annotation.Nullable;
 import org.apache.maven.plugin.MojoExecutionException;
+import org.digitalmediaserver.crowdin.api.FileType;
+
 
 /**
  * An abstract {@link org.apache.maven.plugin.Mojo} configuration class
