+++ src/main/java/org/digitalmediaserver/crowdin/configuration/TranslationFileSet.java
@@ -27,6 +27,7 @@ import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import org.apache.maven.plugin.Mojo;
 import org.apache.maven.plugin.MojoExecutionException;
+import org.digitalmediaserver.crowdin.api.FileType;
 import org.digitalmediaserver.crowdin.tool.CrowdinFileSystem;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
